// development/OnBoarding/OnBoarding/OnBoarding.Web.Fnz/src/typings/ApiClient.ts
export class AssociatePlan implements IAssociatePlan {
    id!: string;
    associateId!: string;
    associateCode?: string | undefined;
    associateName?: string | undefined;
    ownerName?: string | undefined;
    modeName?: string | undefined;
    reference?: string | undefined;
    knowledgeTransferTitle?: string | undefined;
    duration!: number;
    week!: number;
    day!: number;
    scheduledDate!: Date;
    completionDate?: Date | undefined;
    proof?: string | undefined;
    proofType!: number;

    constructor(data?: IAssociatePlan) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.associateId = data["AssociateId"];
            this.associateCode = data["AssociateCode"];
            this.associateName = data["AssociateName"];
            this.ownerName = data["OwnerName"];
            this.modeName = data["ModeName"];
            this.reference = data["Reference"];
            this.knowledgeTransferTitle = data["KnowledgeTransferTitle"];
            this.duration = data["Duration"];
            this.week = data["Week"];
            this.day = data["Day"];
            this.scheduledDate = data["ScheduledDate"] ? new Date(data["ScheduledDate"].toString()) : <any>undefined;
            this.completionDate = data["CompletionDate"] ? new Date(data["CompletionDate"].toString()) : <any>undefined;
            this.proof = data["Proof"];
            this.proofType = data["ProofType"];
        }
    }

    static fromJS(data: any): AssociatePlan {
        data = typeof data === 'object' ? data : {};
        let result = new AssociatePlan();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["AssociateId"] = this.associateId;
        data["AssociateCode"] = this.associateCode;
        data["AssociateName"] = this.associateName;
        data["OwnerName"] = this.ownerName;
        data["ModeName"] = this.modeName;
        data["Reference"] = this.reference;
        data["KnowledgeTransferTitle"] = this.knowledgeTransferTitle;
        data["Duration"] = this.duration;
        data["Week"] = this.week;
        data["Day"] = this.day;
        data["ScheduledDate"] = this.scheduledDate ? this.scheduledDate.toISOString() : <any>undefined;
        data["CompletionDate"] = this.completionDate ? this.completionDate.toISOString() : <any>undefined;
        data["Proof"] = this.proof;
        data["ProofType"] = this.proofType;
        return data; 
    }
}

export interface IAssociatePlan {
    id: string;
    associateId: string;
    associateCode?: string | undefined;
    associateName?: string | undefined;
    ownerName?: string | undefined;
    modeName?: string | undefined;
    reference?: string | undefined;
    knowledgeTransferTitle?: string | undefined;
    duration: number;
    week: number;
    day: number;
    scheduledDate: Date;
    completionDate?: Date | undefined;
    proof?: string | undefined;
    proofType: number;
}
// ...
export class KnowledgeTransfer implements IKnowledgeTransfer {
    id!: string;
    name?: string | undefined;
    description?: string | undefined;
    modeId!: string;
    ownerId?: string | undefined;
    reference?: string | undefined;
    duration!: number;

    constructor(data?: IKnowledgeTransfer) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.name = data["Name"];
            this.description = data["Description"];
            this.modeId = data["ModeId"];
            this.ownerId = data["OwnerId"];
            this.reference = data["Reference"];
            this.duration = data["Duration"];
        }
    }

    static fromJS(data: any): KnowledgeTransfer {
        data = typeof data === 'object' ? data : {};
        let result = new KnowledgeTransfer();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Name"] = this.name;
        data["Description"] = this.description;
        data["ModeId"] = this.modeId;
        data["OwnerId"] = this.ownerId;
        data["Reference"] = this.reference;
        data["Duration"] = this.duration;
        return data; 
    }
}

export interface IKnowledgeTransfer {
    id: string;
    name?: string | undefined;
    description?: string | undefined;
    modeId: string;
    ownerId?: string | undefined;
    reference?: string | undefined;
    duration: number;
}
// ...
export class Project implements IProject {
    id!: string;
    code?: string | undefined;
    name?: string | undefined;

    constructor(data?: IProject) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.code = data["Code"];
            this.name = data["Name"];
        }
    }

    static fromJS(data: any): Project {
        data = typeof data === 'object' ? data : {};
        let result = new Project();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Code"] = this.code;
        data["Name"] = this.name;
        return data; 
    }
}

export interface IProject {
    id: string;
    code?: string | undefined;
    name?: string | undefined;
}
// ...
export class ProjectPlan implements IProjectPlan {
    id!: string;
    roleId!: string;
    projectId!: string;
    knowledgeTransferId!: string;
    week!: number;
    day!: number;
    knowledgeTransfer?: KnowledgeTransfer | undefined;
    associatePlan?: AssociatePlan[] | undefined;

    constructor(data?: IProjectPlan) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.roleId = data["RoleId"];
            this.projectId = data["ProjectId"];
            this.knowledgeTransferId = data["KnowledgeTransferId"];
            this.week = data["Week"];
            this.day = data["Day"];
            this.knowledgeTransfer = data["KnowledgeTransfer"] ? KnowledgeTransfer.fromJS(data["KnowledgeTransfer"]) : <any>undefined;
            if (data["AssociatePlan"] && data["AssociatePlan"].constructor === Array) {
                this.associatePlan = [] as any;
                for (let item of data["AssociatePlan"])
                    this.associatePlan!.push(AssociatePlan.fromJS(item));
            }
        }
    }

    static fromJS(data: any): ProjectPlan {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectPlan();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["RoleId"] = this.roleId;
        data["ProjectId"] = this.projectId;
        data["KnowledgeTransferId"] = this.knowledgeTransferId;
        data["Week"] = this.week;
        data["Day"] = this.day;
        data["KnowledgeTransfer"] = this.knowledgeTransfer ? this.knowledgeTransfer.toJSON() : <any>undefined;
        if (this.associatePlan && this.associatePlan.constructor === Array) {
            data["AssociatePlan"] = [];
            for (let item of this.associatePlan)
                data["AssociatePlan"].push(item.toJSON());
        }
        return data; 
    }
}

export interface IProjectPlan {
    id: string;
    roleId: string;
    projectId: string;
    knowledgeTransferId: string;
    week: number;
    day: number;
    knowledgeTransfer?: KnowledgeTransfer | undefined;
    associatePlan?: AssociatePlan[] | undefined;
}
// ...
export class Role implements IRole {
    id!: string;
    code?: string | undefined;
    name?: string | undefined;

    constructor(data?: IRole) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["Id"];
            this.code = data["Code"];
            this.name = data["Name"];
        }
    }

    static fromJS(data: any): Role {
        data = typeof data === 'object' ? data : {};
        let result = new Role();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["Id"] = this.id;
        data["Code"] = this.code;
        data["Name"] = this.name;
        return data; 
    }
}

export interface IRole {
    id: string;
    code?: string | undefined;
    name?: string | undefined;
}